refactor(pack): refactor bdhcp
