+++ packages/manager/apps/telecom/src/app/app.less
@@ -108,19 +108,6 @@
 @import './telecom/pack/xdsl/access/pack-xdsl-access.less';
 @import './telecom/pack/xdsl/modem/pack-xdsl-modem.less';
 @import './telecom/pack/xdsl/orderFollowUp/orderFollowUp.less';
-@import './telecom/telephony/alias/portability/portabilities/telecom-telephony-alias-portability-portabilities.less';
-@import './telecom/telephony/billingAccount/billing/groupRepayments/telecom-telephony-billing-account-billing-group-repayments.less';
-@import './telecom/telephony/billingAccount/dashboard/telecom-telephony-billing-account-dashboard.less';
-@import './telecom/telephony/billingAccount/manageContacts/telecom-telephony-billing-account-manageContacts.less';
-@import './telecom/telephony/billingAccount/orderAlias/telecom-telephony-billing-account-orderAlias.less';
-@import './telecom/telephony/line/assist/rma/telecom-telephony-line-assist-rma.less';
-@import './telecom/telephony/line/assist/troubleshooting/telecom-telephony-line-assist-troubleshooting.less';
-@import './telecom/telephony/line/calls/click2Call/telecom-telephony-line-calls-click2Call.less';
-@import './telecom/telephony/line/phone/accessories/accessories.less';
-@import './telecom/telephony/line/phone/configuration/telecom-telephony-line-phone-configuration.less';
-@import './telecom/telephony/service/fax/campaigns/read/telecom-telephony-service-fax-campaigns-read.less';
-@import './telecom/telephony/service/fax/settings/telecom-telephony-service-fax-settings.less';
-@import './telecom/telephony/telecom-telephony.less';
 
 @import './telecom/telephony/line/phone/order/choice/choice.less';
 @import './telecom/pack/meetings/xdsl-meetings.less';
